fix: ranking retornava 500 - ROUND(double precision, n) inexistente no Postgres, usar numeric

+++ backend/src/routes/competitions.js
@@ -176,7 +176,7 @@ router.get('/last-finished', auth, async (req, res) => {
           ELSE 0
         END as media,
         CASE WHEN cp.total_answered > 0
-          THEN ROUND((cp.correct_answers::float / cp.total_answered) * 100, 1)
+          THEN ROUND((cp.correct_answers::numeric / cp.total_answered) * 100, 1)::float
           ELSE 0
         END as accuracy
       FROM competition_participants cp
